Data/requirements.txt
camelot-py==1.0.0
pandas==2.2.3
pytz==2025.1
